+++ src/endstone/core/server.h
@@ -46,6 +46,7 @@ class EndstoneIpBanList;
 class EndstoneMetrics;
 class EndstonePlayer;
 class EndstonePlayerBanList;
+class PluginMetricsRegistry;
 class EndstoneServer : public Server {
 public:
     explicit EndstoneServer();
@@ -75,7 +76,8 @@ public:
 
     [[nodiscard]] Scheduler &getScheduler() const override;
     [[nodiscard]] EndstoneScheduler &getEndstoneScheduler() const;
-    [[nodiscard]] std::unique_ptr<MetricsBase> createMetrics(Plugin &plugin, int service_id) const override;
+    [[nodiscard]] NotNull<MetricsBase> createMetrics(Plugin &plugin, int service_id) const override;
+    [[nodiscard]] PluginMetricsRegistry &getPluginMetrics() const { return *plugin_metrics_; }
 
     [[nodiscard]] Level &getLevel() const override;
     [[nodiscard]] EndstoneLevel *getEndstoneLevel() const;
@@ -164,6 +166,7 @@ private:
     std::unordered_map<std::type_index, std::unique_ptr<IRegistry>> registries_;
     Nullable<EndstoneScoreboard> scoreboard_;
     std::unique_ptr<EndstoneMetrics> metrics_;
+    std::unique_ptr<PluginMetricsRegistry> plugin_metrics_;
     std::unordered_map<UUID, NotNull<EndstoneScoreboard>> player_boards_;
     std::chrono::system_clock::time_point start_time_;
     IResourcePackRepository *resource_pack_repository_ = nullptr;